+++ iniparser/dictionary.h
@@ -85,7 +85,7 @@ unsigned dictionary_hash(char * key);
   dictionary, give size=0.
  */
 /*--------------------------------------------------------------------------*/
-dictionary * dictionary_new(int size);
+dictionary * dictionary_new(size_t size);
 
 /*-------------------------------------------------------------------------*/
 /**
